Refactor app configs to own file
Ref oursky/likecoin-chain#119

Based on https://github.com/nnkken/likecoin-chain/tree/v0.44-dual-prefix
